spotify-streaming-pipeline/cloud_function/requirements.txt
functions-framework
requests
google-cloud-pubsub
